requirements.txt
faker==40.11.0
pytest==8.4.2
python-dotenv==1.2.2
playwright==1.58.0
pytest-playwright==0.7.2

